requirements.txt
exrex==0.12.0
httpx==0.28.1
PyYAML==6.0.2
httpx[http2]